refactor(tests): split dns resolver suite by concern
Replace the monolithic DNS resolver test module with record, cache, and configuration-focused suites, add shared DNS resolver test support helpers, and preserve existing resolver behavior coverage with explicit error-message assertions.
